// src/redux/actions/types/scheduleScenariosActionTypes.ts
export enum EScheduleScenariosActionTypes {
  LOAD_SCENARIOS = 'ScheduleScenarios/LOAD_SCENARIOS',
  OPEN_SCENARIO_SNAPSHOT = 'ScheduleScenarios/OPEN_SCENARIO_SNAPSHOT',
  SELECT_SCENARIO = 'ScheduleScenarios/SELECT_SCENARIO',
  SET_OPEN_NEW_SCENARIO_WIZARD = 'ScheduleScenarios/SET_OPEN_NEW_SCENARIO_WIZARD',
  SORT_SCENARIOS = 'ScheduleScenarios/SORT_SCENARIOS',
  GET_SITES = 'ScheduleScenarios/GET_SITES',
  SET_SELECTED_SCENARIO_TYPE = 'ScheduleScenarios/SET_SELECTED_SCENARIO_TYPE',
  TOGGLE_DELETE_SCENARIO_POPUP = 'ScheduleScenarios/TOGGLE_DELETE_SCENARIO_POPUP',
  DELETE_SCENARIO = 'ScheduleScenarios/DELETE_SCENARIO',
  LOAD_FORECAST_SCENARIOS = 'ScheduleScenarios/LOAD_FORECAST_SCENARIOS',
  SAVE_SCENARIO_COMMENTS = 'ScheduleScenarios/SAVE_SCENARIO_COMMENTS',
  RENAME_SCENARIO = 'ScheduleScenarios/RENAME_SCENARIO',
  FIND_SCENARIOS = 'ScheduleScenarios/FIND_SCENARIOS',
  TOGGLE_FILTER_MENU = 'ScheduleScenarios/TOGGLE_FILTER_MENU',
  TOGGLE_FILTER_ITEM = 'ScheduleScenarios/TOGGLE_FILTER_ITEM',
}